+++ spead2/test/test_recv.py
@@ -460,9 +460,8 @@ class TestUdpReader(object):
             stream.add_udp_reader(123)
 
 
-class TestThreads(object):
-    """Tests for situations that could cause deadlocks or other thread safety
-    issues."""
+class TestStream(object):
+    """Tests for the stream API"""
 
     def __init__(self):
         self.flavour = FLAVOUR
@@ -478,7 +477,23 @@ class TestThreads(object):
                     shape=data.shape, dtype=data.dtype, value=data)
         gen = spead2.send.HeapGenerator(ig)
         for i in range(10):
-            sender.send_heap(gen.get_heap())
+            sender.send_heap(gen.get_heap(data='all'))
         receiver = spead2.recv.Stream(thread_pool)
         receiver.add_buffer_reader(sender.getvalue())
         receiver.stop()
+
+    def test_no_stop_heap(self):
+        """A heap containing a stop is not passed to the ring"""
+        thread_pool = spead2.ThreadPool(1)
+        sender = spead2.send.BytesStream(thread_pool)
+        ig = spead2.send.ItemGroup()
+        data = np.array([[6, 7, 8], [10, 11, 12000]], dtype=np.uint16)
+        ig.add_item(id=0x2345, name='name', description='description',
+                    shape=data.shape, dtype=data.dtype, value=data)
+        gen = spead2.send.HeapGenerator(ig)
+        sender.send_heap(gen.get_heap(data='all'))
+        sender.send_heap(gen.get_end())
+        receiver = spead2.recv.Stream(thread_pool)
+        receiver.add_buffer_reader(sender.getvalue())
+        heaps = list(receiver)
+        assert_equal(1, len(heaps))